requirements.txt
aniso8601==7.0.0
Click==7.0
celery==4.3.0 
Django==2.2.6
django-celery-beat==1.5.0
django-filter==2.2.0
django-model-utils==3.2.0
Flask==1.1.1
Flask-GraphQL==2.0.0
Flask-SQLAlchemy==2.4.1
flower==0.9.3
graphene==2.1.8
graphene-django==2.6.0
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
numpy==1.17.2
pandas==0.25.1
promise==2.2.1
pyasn1==0.4.7
pyexasol==0.8.0
pyodbc==4.0.27
python-dateutil==2.8.0
pytz==2019.2
rsa==4.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.7
sqlalchemy-exasol==2.0.6
sqlparse==0.3.0
websocket-client==0.56.0
Werkzeug==0.16.0
